t5548: new test cases for push --porcelain and --dry-run
Refactor t5548 and add new test cases for git-push with both --porcelain
and --dry-run flags in order to cover the issues reported by Patrick.

When executing git-push(1) with the "--porcelain" flag, then we will
print updated references in a machine-readable format that looks like
this:

To destination
=   refs/heads/noop:refs/heads/noop [up to date]
!   refs/heads/rejected:refs/heads/rejected [rejected] (atomic push failed)
!   refs/heads/noff:refs/heads/(off (non-fast-forward)
Done

The final "Done" stanza was introduced via 77555854be (git-push: make
git push --porcelain print "Done", 2010-02-26), where it was printed
"unless any errors have occurred". For the purpose of the "Done" line,
knowing a ref will be rejected in a --dry-run does not count as an
error. Actual rejections in non --dry-run pushes do count as errors.

The new test cases will be used in the next commit to protect the above
behaviors of porcelain output of git push.

Reported-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>