push: only ignore finish_connect() for dry-run mode
Patrick reported an issue that the exit code of git-receive-pack(1) is
ignored during atomic push with "--porcelain" flag, and added new test
cases in t5543.

Atomic push may abort the connection early and close the pipe, which
may cause an error for `finish_connect()`. Arbitrarily ignoring this
error will lead to issues. Modify it to only ignore the finish_connect()
error when both the --atomic and --dry-run flags are set.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>